Solve day 6 Psrt 2

+++ aoc2023/src/main/kotlin/com/adventofcode/samour/aoc2023/day6/RaceRecord.kt
@@ -1,6 +1,8 @@
 package com.adventofcode.samour.aoc2023.day6
 
+import java.math.BigInteger
+
 data class RaceRecord(
-    val raceTime: Int,
-    val recordDistance: Int,
+    val raceTime: BigInteger,
+    val recordDistance: BigInteger,
 )
